[CMAKE] optional build components
For library developer that may include this project as subdirectory it is best practice to disable tests by default and offer an option to disable executables. Therefore the cmake option 'AMITY_BUILD_EXECUTABLES' can be set to off for skipping add_executable and 'AMITY_BUILD_CRYPTOTEST' must be turned on in order to run the test suite.