[CMAKE] conditional c++17 flag for msvc
MSVC throws errors on c++17 features like '[[fallthrough]]' even if the compiler is capable of but the corresponding flag is not set. Thus if we are compiling using msvc and c++17 is supported we set the corresponding flag. Note: The c++17 features used should be encapsulated if c++17 is not required using #ifdef-soup otherwise the CMAKE_CXX_STANDARD and CMAKE_CXX_STANDRAD_REQUIRED cmake variables should be set.