[CMAKE] fixed versioning
First the git commit versioning was broken and i reimplemented all of it to be shorter more robust and more convenient. Second the version file always changed in every build forcing to recompile several files. The version file is now configured on the fly using the version target and checks if anything changes at alle before replacing the old one.