[CI] appveyor pipeline: GitHub Releases for debain and windows (#8)
* Adding brandonlehmann's checkpointer scripts

* Updating version

* [CMAKE] commit id version option

The actual version script is bugged for general builds and should only be used for production/staging deployments. Thus you are now able to set the cmake option `AMITY_ENABLE_GIT_VERSION_ID` in your ci deployment.

* [CMAKE] optional build components

For library developer that may include this project as subdirectory it is best practice to disable tests by default and offer an option to disable executables. Therefore the cmake option 'AMITY_BUILD_EXECUTABLES' can be set to off for skipping add_executable and 'AMITY_BUILD_CRYPTOTEST' must be turned on in order to run the test suite.

* [CMAKE] install targets for executables

In order to install the applications `cmake --build . --target install` is now sufficient. Further the install targets for the miniupnpc dependency are removed such that users are not bloated with unwanted/needed external libraries.

* [GIT] ignore build artifacts

* [CI] initial build script

* [CMAKE] conditional c++17 flag for msvc

MSVC throws errors on c++17 features like '[[fallthrough]]' even if the compiler is capable of but the corresponding flag is not set. Thus if we are compiling using msvc and c++17 is supported we set the corresponding flag. Note: The c++17 features used should be encapsulated if c++17 is not required using #ifdef-soup otherwise the CMAKE_CXX_STANDARD and CMAKE_CXX_STANDRAD_REQUIRED cmake variables should be set.

* [CMAKE] bugfix missing target version

If 'AMITY_ENABLE_GIT_VERSION_ID' is set OFF the version target had not been created at all throwing cmake errors. Therefore a proxy target doing nothing fixes the issue.

* [CODE] removed single c++17 part

since the code isn't buildable using the newest standard c++17 features and only one part of the code uses one c++17 feature the code has been slightly transformed such that c++17 is not required.

* [CI] removed cmake cache removal

since appveyor starts every build process in a fresh vm, with same setup, hard wireing shouldnt be an issue. Therefore deleting the cmake cache should not be necessary.

* [CI] dropped debug configuration

Appveyor takes far to long for the build process, because debug build is just a compilation check it may can be neglected for now

* [CI] artifacts upload

* [CI] install location fix

When installing using `cmake --build . --target install` the install location seems bugged if the prefix provided in configuration was untyped.

* [CI] added github deployment

* [CI] applied os system to artifact name

If the artifact name is identical appveyor overwrites the previous one, therefor they must be image specific.

* Update hash.h

Comment correction

* [CI] removed cache

Since the cache is not used on recompilation it brings an overhead downloading unzipping and zipping it.

* [CI] deploy releases to github only tagged master

* Update hash.h

Correction to CN_SOFT_SHELL_MULTIPLIER

* [CMAKE] enabling latest c++ standard if using msvc

* [CMAKE] refactored c++ compiler flags for msvc

* [CODE] removed std::unitary_function

is deprecetaded and removed in c++17, it is sufficient to delete the inherited typedef in favor to `-> auto`, `declval` aso.

* [CODE] removed std::random_shuffle

random_shuffle is removed and std::shuffle now takes a random generator as an argument. Therefore I changed the shuffle algorithms such that the same generator as before is used.

* [CMAKE] c++17 support for gcc

* [CMAKE] updated required cmake version to 3.9.2

* [CMAKE] moved default build type into a seperated file

* [CMAKE] fixed versioning

First the git commit versioning was broken and i reimplemented all of it to be shorter more robust and more convenient. Second the version file always changed in every build forcing to recompile several files. The version file is now configured on the fly using the version target and checks if anything changes at alle before replacing the old one.

* [CMAKE] compiler setup seperation

The main CMakeLists.txt was pretty much bloated. I seperated configurations and moved them to a shared configuration in `./cmake/*.cmake` or into the corresponding subdirectory `./src` and `./external`.

* [CI] applied the version file

The version file `./VERSION` will now be read by the pipeline and artifacts will be named correctly.

* [CI] appveyor discord notification

* [CMAKE] fixed compiler warnings/error

The external projects did override some compiler flags leading to warnings using msvc.

* [CODE] renamed turtlecoind to amitycoind

* [CMAKE] [CI] fixed branch versioning name in appveyor

* [CMAKE] fixed linux building issue

the rocksdb library was not setup correctly due to a new introduced subdirectory resulting in `external/external/rockdb`

* [CODE] integration tests for validating hash consistency