[RPC] [FIX] connection error handling
The init method of NodeRpcProxy always switched to INITIALIZED even if it couldnt connect leading to segmentation faults because users of the node thought its online and made calls even if they checked the state before. Therefore I an additional check and resetted the state to NOT_INITAILIZED once the worker thread shutdowns.